fix: crashing when returning from route detail
this is so fucked i dont want to talk about it
